+++ entity/src/epilot/models/operations/getentityactivityfeed.py
@@ -23,10 +23,12 @@ class GetEntityActivityFeedRequest:
     r"""get activities before this timestamp"""
     from_: Optional[int] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'from', 'style': 'form', 'explode': True }})
     r"""start from page"""
+    include_relations: Optional[bool] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'include_relations', 'style': 'form', 'explode': True }})
+    r"""Include activities from related entities"""
     size: Optional[int] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'size', 'style': 'form', 'explode': True }})
     r"""max number of results to return"""
     type: Optional[str] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'type', 'style': 'form', 'explode': True }})
-    r"""Activity type"""
+    r"""Filter by activity type"""