+++ entity/src/epilot/saved_views.py
@@ -94,6 +94,34 @@ class SavedViews:
         return res
 
     
+    def list_favorite_views_for_user(self) -> operations.ListFavoriteViewsForUserResponse:
+        r"""listFavoriteViewsForUser
+        Get the Favorite Saved Views for user based on the schema
+        """
+        base_url = utils.template_url(*self.sdk_configuration.get_server_details())
+        
+        url = base_url + '/v1/entity/views/favorites'
+        headers = {}
+        headers['Accept'] = 'application/json'
+        headers['user-agent'] = f'speakeasy-sdk/{self.sdk_configuration.language} {self.sdk_configuration.sdk_version} {self.sdk_configuration.gen_version} {self.sdk_configuration.openapi_doc_version}'
+        
+        client = self.sdk_configuration.security_client
+        
+        http_res = client.request('GET', url, headers=headers)
+        content_type = http_res.headers.get('Content-Type')
+
+        res = operations.ListFavoriteViewsForUserResponse(status_code=http_res.status_code, content_type=content_type, raw_response=http_res)
+        
+        if http_res.status_code == 200:
+            if utils.match_content_type(content_type, 'application/json'):
+                out = utils.unmarshal_json(http_res.text, Optional[operations.ListFavoriteViewsForUser200ApplicationJSON])
+                res.list_favorite_views_for_user_200_application_json_object = out
+            else:
+                raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
+
+        return res
+
+    
     def list_saved_views(self) -> operations.ListSavedViewsResponse:
         r"""listSavedViews
         Get the Saved Views based on the schema